Fix WCF reconnection, change TCP connection to pipes, fix some issues (removing job, opening a dialog in non-blocked mode for the service).

+++ AutoPrintr/AutoPrintr.Core/Services/FileService.cs
@@ -31,6 +31,11 @@ namespace AutoPrintr.Core.Services
             {
                 try
                 {
+                    if (!File.Exists(filePath))
+                    {
+                        return false;
+                    }
+
                     File.SetAttributes(filePath, FileAttributes.Normal);
                     File.Delete(filePath);
                     return true;
